+++ pkg/local_object_storage/engine/gc_test.go
@@ -182,7 +182,6 @@ func TestGC(t *testing.T) {
 				meta.WithEpochState(es),
 			),
 			shard.WithExpiredObjectsCallback(e.processExpiredObjects),
-			shard.WithExpiredLocksCallback(e.processExpiredLocks),
 			shard.WithGCRemoverSleepInterval(100*time.Millisecond),
 			shard.WithContainerPayments(containerPaymantsStub{}),
 		)